### 2025-02-20 #7805 @editorialbot
[PRE REVIEW]: NPSdataverse: a suite of R packages for data processing, authoring Ecological Metadata Language metadata, checking data-metadata congruence, and accessing data
**Submitting author:** @RobLBaker (<a href="http://orcid.org/0000-0001-7591-5035">Robert L. Baker</a>)
**Repository:** https://github.com/nationalparkservice/NPSdataverse.git
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
