### 2025-02-20 2226a4e
Add JOSS submission badge; update paper.bib
