### 2025-02-20 02ba3f7
Update references, add in DOIs and fix a URL.
