+++ paper.bib
@@ -18,12 +18,12 @@
   }
 
 @manual{nsf2015,
-  url = {https://www.nsf.gov/publications/pub_summ.jsp?ods_key=nsf15094},
+  url = {https://www.nsf.gov/notices/general/national-science-foundation-open-government-plan-40},
   title = {The National Science Foundation Open Government Plan 3.5},
-  howpublished = {NSF Document Number nsf15093},
+  howpublished = {Publication number: NSF 16-131},
   organization = {National Science Foundation},
   address = {Alexandria, VA, USA},
-  year = {2015}
+  year = {2016}
   }
 
 @misc{Wiley2022,
@@ -121,7 +121,7 @@
    volume = "37",
    number = "Volume 37, 2006",
    pages = "519-544",
-   doi = "https://doi.org/10.1146/annurev.ecolsys.37.091305.110031",
+   doi = "10.1146/annurev.ecolsys.37.091305.110031",
    url = "https://www.annualreviews.org/content/journals/10.1146/annurev.ecolsys.37.091305.110031",
    publisher = "Annual Reviews",
    issn = "1545-2069",
@@ -174,9 +174,10 @@ url = {https://doi.org/10.1038/sdata.2016.18},
 @Manual{Smith2022,
   title = {EMLassemblyline: A tool kit for building EML metadata workflows},
   author = {Colin Smith},
-  year = {2022},
-  note = {https://github.com/EDIorg/EMLassemblyline,
-https://ediorg.github.io/EMLassemblyline/},
+  year = {2025},
+  note = {R package version 3.5.7,
+    https://ediorg.github.io/EMLassemblyline/},
+  url = {https://github.com/EDIorg/EMLassemblyline},
 }
 
 @Manual{Baker_QCkit2025,
@@ -184,6 +185,7 @@ https://ediorg.github.io/EMLassemblyline/},
   author = {Baker, Robert and Judd Patterson and Joe DeVivo and Issac Quevedo and Sarah Wright},
   year = {2025},
   note = {R package version 1.0.0},
+  doi = {10.57830/2307726},
   url = {https://github.com/nationalparkservice/QCkit/},
 }
 
@@ -200,6 +202,7 @@ https://ediorg.github.io/EMLassemblyline/},
   author = {Baker, Robert and Judd Patterson},
   year = {2025},
   note = {R package version 1.0.0},
+  doi = {10.57830/2307727},
   url = {https://github.com/nationalparkservice/EMLeditor},
 }
 
@@ -208,6 +211,7 @@ https://ediorg.github.io/EMLassemblyline/},
   author = {Baker, Robert and Sarah E. Wright},
   year = {2025},
   note = {R package version 1.0.0},
+  doi = {10.57830/2307755},
   url = {https://nationalparkservice.github.io/DPchecker/},
 }
 
@@ -216,6 +220,7 @@ https://ediorg.github.io/EMLassemblyline/},
   author = {Baker, Robert and Joe DeVivo and Judd Patterson},
   year = {2025},
   note = {R package version 1.0.0},
+  doi = {10.57830/2307756},
   url = {https://github.com/nationalparkservice/NPSutils},
 }
 
@@ -226,5 +231,6 @@ https://ediorg.github.io/EMLassemblyline/},
   institution = {National Park Service},
   type = {},
   number = {},
-  note = {\url{https://doi.org/10.57830/2300498}}
+  url = {https://doi.org/10.57830/2300498},
+  doi = {10.57830/2300498}
   }
