### 2025-02-20 #7805 @RobLBaker
> **License info**:
>
> 🟡 License found: `Other` ([Check here](https://opensource.org/licenses) for OSI approval)

Creative Commons Zero (CC0) license is a requirement of our federal agency. Unfortunately, I can't change the license. However, I can suggest that you add CC0 license to the list of OSI approved licenses ;). CC0 is the least restrictive license we have been able to identify, other than "public domain" which is not, as far as I can tell a type of license: https://creativecommons.org/public-domain/cc0/
