+++ paper.bib
@@ -63,6 +63,7 @@
   number = {1},
   pages = {192},
   year = {2021},
+  doi = {10.1038/s41597-021-00981-0},
   publisher = {Nature Publishing Group UK London}
 }
 
@@ -85,6 +86,7 @@
   number = {4},
   pages = {e2018},
   year = {2022},
+  doi = {10.1002/bes2.2018},
   publisher = {Wiley Online Library}
 }
 
@@ -93,7 +95,8 @@
   url = {https://eml.ecoinformatics.org}, DOI={10.5063/f11834t2},
   publisher = {KNB Data Repository},
   author = {Jones, Matthew and O'Brien, Margaret and Mecum, Bryce and Boettiger, Carl and Schildhauer, Mark and Maier, Mitchell and Whiteaker, Timothy and Earl, Stevan and Chong, Steven},
-  year = {2019}
+  year = {2019},
+  doi = {10.32614/cran.package.eml}
   }
 
 @article{Nelson2022,
@@ -110,7 +113,8 @@
   volume = {4},
   number = {34},
   pages = {1276},
-  year = {2019}
+  year = {2019},
+  doi = {10.21105/joss.01276}
 }
 
 @article{Jones2006,
