### 2025-02-20 #7805 @RobLBaker
checked for "skip DOIs" and added when appropriate.
Added missing DOIs.
Fixed invalid DOIs.
