### 2025-02-21 #7805 @Kevin-Mattheus-Moerman
- [ ] The project needs to carry an OSI approved license. The CC0 license is not acceptable. You can still use CCO for data, content, documents etc if you like, however the software should carry an OSI approved license. Can you send any link to documentation from your federal agency on licenses? I wonder if they actually mean the above.
- [ ] Can you clarify exactly what it is you are seeking to publish in JOSS? Your software seems to gather the functionality of other packages in about 149 lines of R. Are you the author of all those packages and you seek the joint publication of the entire suite? Or are you only the author of this submission? If it is the latter than this work is not acceptable for JOSS. First of all it is too small and secondly it does not offer significantly added functionality.
