### 2025-02-21 dd111ab
Add a file outlining how to contribute to the NPSdataverse package
