+++ CONTRIBUTING.md
@@ -38,7 +38,7 @@ If you then still feel the need to ask a question and need clarification, we rec
 
 - Open an [Issue](https://github.com/nationalparkservice/NPSdataverse/issues/new/choose).
 - Provide as much context as you can about what you're running into.
-- Provide project and platform versions (nodejs, npm, etc), depending on what seems relevant.
+- Provide project and platform versions , depending on what seems relevant.
 
 We will then take care of the issue as soon as possible.
 
@@ -77,7 +77,7 @@ We use GitHub issues to track bugs and errors. If you run into an issue with the
 
 - Open an [Issue](https://github.com/nationalparkservice/NPSdataverse/issues/new/choose). (Since we can't be sure at this point whether it is a bug or not, we ask you not to talk about a bug yet and not to label the issue.)
 - Explain the behavior you would expect and the actual behavior.
-- Please provide as much context as possible and describe the *reproduction steps* that someone else can follow to recreate the issue on their own. This usually includes your code. For good bug reports you should isolate the problem and create a reduced test case.
+- Please provide as much context as possible and describe the *reproduction steps* that someone else can follow to recreate the issue on their own. This usually includes your code. For good bug reports you should isolate the problem and create a reduced test case. See [https://gist.github.com/hadley/270442](https://gist.github.com/hadley/270442) for a great run down on how to create reproducible examples.
 - Provide the information you collected in the previous section.
 
 Once it's filed:
@@ -96,7 +96,8 @@ This section guides you through submitting an enhancement **including completely
 - Make sure that you are using the latest version.
 - Read the [documentation](https://nationalparkservice.github.io/NPSdataverse/) carefully and find out if the functionality is already covered, maybe by an individual configuration.
 - Perform a [search](https://github.com/nationalparkservice/NPSdataverse/issues) to see if the enhancement has already been suggested. If it has, add a comment to the existing issue instead of opening a new one.
-- Find out whether your idea fits with the scope and aims of the project. It's up to you to make a strong case to convince the project's developers of the merits of this feature. Keep in mind that we want features that will be useful to the majority of our users and not just a small subset. If you're just targeting a minority of users, consider writing an add-on/plugin library.
+- Find out whether your idea fits with the scope and aims of the project. It's up to you to make a strong case to convince the project's developers of the merits of this feature. Keep in mind that we want features that will be useful to the majority of our users and not just a small subset. If you're just targeting a minority of users, consider writing an add-on/plugin library or package.
+- Please consult with the developers if you would like to create an add-on package. Maybe we can help!
 
 
 #### How Do I Submit a Good Enhancement Suggestion?
