### 2025-03-10 #7805 @RobLBaker
@Kevin-Mattheus-Moerman I believe I have resolved two of your three questions above.  If I have not resolved the CONTRIBUTING.md as well as number of packages questions to your satisfaction, could you please let me know what additional information I need to supply?

I seek additional further clarification and advice on the licensing issue.  I have received approval to use a non-public domain and non-CC0 license provided that it confers similar restrictions (i.e. none whatsoever).

The closest license I have found that is OSI-approved is the "Zero-Clause BSD" (0BSD): https://opensource.org/license/0bsd

However, this is not a standard license for R software and causes warnings upon conducting local build checks. While normally I am not adverse to a warning if the cause is known, local build check warnings result in failure for the github action build checks.

Can you advise on the best way to proceed? Do you know how other US government agencies have handled JOSS publications? Is there another OSI-approved license that is similar to 0BSD? I have not yet found one. Is there an acceptable way to include a non-standard license in an R package and avoid the build check warnings? I have not been able to find this either.
