### 2025-03-12 92d8fd0
update license with usethis::use_mit_license
