### 2025-03-12 a861a66
Add info to changelog about MIT license
