+++ NEWS.md
@@ -1,4 +1,6 @@
 # NPSdataverse 1.0.1 (development version)
+2025-03-12
+* Updated to MIT license, which is OSI approved and NPS approved and a standard R license.
 2025-03-10
 * Updated license to OSI-approved "Zero-Clause BSD" in support of JOSS submission.
 
