### 2025-03-13 #7805 @RobLBaker
@Kevin-Mattheus-Moerman I believe I have sorted out the licensing issue. I got permission to use the MIT license which is OSI-approved and has been applied to all the NPSdataverse meta-package as well as all of the constituent packages that make up the submission.

Can you please advise if there are any other steps I need to facilitate advancing the submission to the review process?
