### 2025-04-07 #7805 @goldingn
Hi @RobLBaker. Thanks for your submission and your patience with the process so far!

Just so you know this is my first article as an editor (though I've authored and reviewed with JOSS before), so there may be some clunkiness as I familiarise myself with the processes and editorialbot commands.

It may also take a little time to get the reviews given the size of the submission; ie. my interpretation of the above is that the software submitted comprises all six component packages, and the umbrella package. But we can make a start and keep in touch to keep things ticking along.
