### 2025-04-07 #7805 @goldingn
Hi @stephpenn1 and @wiljnich. Would either or both of you be interested in reviewing this short JOSS paper - and associated software - on a collection of R packages for working with EML (Ecological Metadata Language)?
