### 2025-04-07 #7805 @stephpenn1
Hi @goldingn, yes, I would be happy to review this JOSS paper
