+++ R/updateR.R
@@ -5,48 +5,55 @@
 #' @return dataframe
 #' @keywords internal
 #'
-.update_git_repos<-function(){
-  git_pkgs = c("NPSdataverse",
-               "QCkit",
-               "EMLeditor",
-               "DPchecker",
-               "NPSutils",
-               "EMLassemblyline")
+.update_git_repos <- function() {
+  git_pkgs <- c(
+    "NPSdataverse",
+    "QCkit",
+    "EMLeditor",
+    "DPchecker",
+    "NPSutils",
+    "EMLassemblyline"
+  )
 
-  pkg_update <- remotes::package_deps(git_pkgs, dependencies = c("Imports",
-                                                                 "Remotes",
-                                                                 "Suggests"))
+  pkg_update <- remotes::package_deps(git_pkgs, dependencies = c(
+    "Imports",
+    "Remotes",
+    "Suggests"
+  ))
 
-  if(any(pkg_update$diff < 0)){
+  if (any(pkg_update$diff < 0)) {
     old_pkgs <- NULL
     old_users <- NULL
-    for(i in seq_along(pkg_update$diff)){
-      if(pkg_update$diff[i] < 0){
+    for (i in seq_along(pkg_update$diff)) {
+      if (pkg_update$diff[i] < 0) {
         old_pkgs <- append(old_pkgs, pkg_update$package[i])
         old_users <- append(old_users, pkg_update$remote[[i]]$username)
       }
     }
     load_header <- cli::rule(
       left = cli::pluralize(
-        "The following {cli::qty(length(old_pkgs))}package{?s} {?is/are} out of date:\n"))
+        "The following {cli::qty(length(old_pkgs))}package{?s} {?is/are} out of date:\n"
+      )
+    )
     msg(load_header)
     .print_cust_package_deps(pkg_update)
     cli::cat_line()
     cli::cli_text(
-      "{.strong To update {cli::qty(length(old_pkgs))}th{?is/ese} {cli::qty(length(old_pkgs))}package{?s}, please run:\n}")
+      "{.strong To update {cli::qty(length(old_pkgs))}th{?is/ese} {cli::qty(length(old_pkgs))}package{?s}, please run:\n}"
+    )
     cli::cat_line("detach_NPSdataverse()")
     cli::cat_line("devtools::install_github(\"", old_users, "/", old_pkgs, "\")")
     cli::cat_line("\nClose R and Rstudio. Open a new R session and reload the NPSdataverse.")
     cli::cat_line()
-  }
-  else{
+  } else {
     load_header <- cli::rule(
       left = cli::pluralize(
-       "All NPSdataverse packages are up to date."))
+        "All NPSdataverse packages are up to date."
+      )
+    )
     msg(load_header)
     cli::cat_line()
   }
-
 }
 
 #' Custom print function for github repos to update
@@ -61,7 +68,7 @@
 #' @return printed text to console
 #' @keywords internal
 #'
-.print_cust_package_deps<-function (x, show_ok = FALSE, ...){
+.print_cust_package_deps <- function(x, show_ok = FALSE, ...) {
   class(x) <- "data.frame"
   x$remote <- lapply(x$remote, format)
   ahead <- x$diff > 0L
@@ -70,7 +77,7 @@
   x$diff <- NULL
   x[] <- lapply(x, remotes:::format_str, width = 12)
   if (any(behind)) {
-    #cat("Needs update -----------------------------\n")
+    # cat("Needs update -----------------------------\n")
     print(x[behind, , drop = FALSE], row.names = FALSE, right = FALSE)
   }
   if (any(ahead)) {
@@ -79,7 +86,9 @@
   }
   if (show_ok && any(same_ver)) {
[…]
