### 2025-04-07 #7805 @RobLBaker
@goldingn Sure thing! Here are the packages:

- [NPSdataverse](https://github.com/nationalparkservice/NPSdataverse): a wrapper/meta-package that loads the following packages (and checks for any updates):

  - [QCkit](https://github.com/nationalparkservice/QCkit): For data Quality Control, suggesting darwin core naming, etc. Also includes a template script for generating Data Release Reports.
  - [EML](https://github.com/ropensci/EML): A package for generating Ecological Metadata
  - [EMLassemblyline](https://github.com/EDIorg/emlAssemblyLine): for automatically ingesting data to generate EML
  - [EMLeditor](https://github.com/nationalparkservice/EMLeditor): for Editing EML and uploading data packages to NPS' DataStore. Also includes a template script for generating data packages.
  - [DPchecker](https://github.com/nationalparkservice/DPchecker): for checking EML/data congruence, some FAIR data checks, and checking for compliance with NPS' DataStore
  - [NPSutils](https://github.com/nationalparkservice/NPSutils): For downloading data and metadata from NPS DataStore; using metadata to call data types when loading data into R
