### 2025-04-08 43a3fd2
Update min required versions Require at least 1.0.0 for NPSdataverse packages, and at least 3.5.7 for EMLassemblyline
