### 2025-04-11 #7805 @RobLBaker
> I'm currently contacting colleagues via other channels to find another reviewer experience with R/EML/related topics. I'll update on progress here soon.

I might suggest njlyon0, angelchen7, brunj7, or rfbrown.  They aren't currently listed as people who have reviewed for JOSS in the past, but they have a lot of experience with Ecological Metadata Language metadata, data package creation/publication and/or R development. Found through a google search so no conflict of interest that I know of.
