### 2025-04-11 958dda0
Remove crayon dependency
