### 2025-04-11 1a1bf7e
fix URL in latest_github_version
