### 2025-04-12 7955cb6
Fix how we were checking for out of date github packages, and clean up the cli message about it
