+++ R/updateR.R
@@ -6,43 +6,55 @@
 #' @keywords internal
 #'
 .update_git_repos <- function() {
-  git_pkgs <- c(
-    "NPSdataverse",
-    "QCkit",
-    "EMLeditor",
-    "DPchecker",
-    "NPSutils",
-    "EMLassemblyline"
+  git_pkgs <- tibble::tibble(
+    package = c(
+      "QCkit",
+      "EMLeditor",
+      "DPchecker",
+      "NPSutils",
+      "EMLassemblyline",
+      "NPSdataverse"
+    ),
+    repo = c(
+      "nationalparkservice",
+      "nationalparkservice",
+      "nationalparkservice",
+      "nationalparkservice",
+      "EDIorg",
+      "nationalparkservice"),
+    branch = c(
+      "master",
+      "main",
+      "main",
+      "master",
+      "main",
+      "main"
+    )
   )
 
-  pkg_update <- remotes::package_deps(git_pkgs, dependencies = c(
-    "Imports",
-    "Remotes",
-    "Suggests"
-  ))
+  git_pkgs$local_version <- lapply(git_pkgs$package, packageVersion)
+  git_pkgs$latest_version <- mapply(function(x, y, z) .latest_github_version(repo = x, repo_owner = y, branch = z), git_pkgs$package, git_pkgs$repo, git_pkgs$branch, USE.NAMES = FALSE, SIMPLIFY = FALSE)
+  git_pkgs$behind <- mapply(function(x, y) x < y, git_pkgs$local_version, git_pkgs$latest_version)
+  git_pkgs$ahead <- mapply(function(x, y) x > y, git_pkgs$local_version, git_pkgs$latest_version)
+  git_pkgs$local_version <- sapply(git_pkgs$local_version, as.character)
+  git_pkgs$latest_version <- sapply(git_pkgs$latest_version, as.character)
 
-  if (any(pkg_update$diff < 0)) {
-    old_pkgs <- NULL
-    old_users <- NULL
-    for (i in seq_along(pkg_update$diff)) {
-      if (pkg_update$diff[i] < 0) {
-        old_pkgs <- append(old_pkgs, pkg_update$package[i])
-        old_users <- append(old_users, pkg_update$remote[[i]]$username)
-      }
-    }
+  if (any(git_pkgs$behind)) {
+    old_pkgs <- git_pkgs[git_pkgs$behind, ]
+
     load_header <- cli::rule(
       left = cli::pluralize(
-        "The following {cli::qty(length(old_pkgs))}package{?s} {?is/are} out of date:\n"
+        "The following {cli::qty(length(old_pkgs$package))}package{?s} {?is/are} out of date:\n"
       )
     )
     packageStartupMessage(load_header)
-    .print_cust_package_deps(pkg_update)
+    packageStartupMessage(.print_cust_package_deps(old_pkgs))
     cli::cat_line()
     cli::cli_text(
-      "{.strong To update {cli::qty(length(old_pkgs))}th{?is/ese} {cli::qty(length(old_pkgs))}package{?s}, please run:\n}"
+      "{.strong To update {cli::qty(length(old_pkgs$package))}th{?is/ese} {cli::qty(length(old_pkgs$package))}package{?s}, please run:\n}"
     )
     cli::cat_line("detach_NPSdataverse()")
-    cli::cat_line("devtools::install_github(\"", old_users, "/", old_pkgs, "\")")
+    cli::cat_line("devtools::install_github(\"", old_pkgs$repo, "/", old_pkgs$package, "\")")
     cli::cat_line("\nClose R and Rstudio. Open a new R session and reload the NPSdataverse.")
     cli::cat_line()
   } else {
[…]
