### 2025-04-16 #7805 @goldingn
Hi @stevekambouris I'm not sure if you got my email, but would you be interested in reviewing this short JOSS paper - and associated software - on a collection of R packages for working with EML (Ecological Metadata Language)?
