### 2025-04-16 #7805 @angelchen7
Hello @goldingn , I'm interested as well and since I've never reviewed for JOSS before, I would like to know how the process works. Thank you!
