### 2025-04-17 #7805 @goldingn
Thanks @njlyon0 and @angelchen7! We have a reviewer guide here with all the details: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Essentially check through the paper (<1000 words), and review the software following a checklist (here's an example: https://github.com/openjournals/joss-reviews/issues/8032#issuecomment-2791829817)

We'll start a new GitHub issue for the main review (this is the pre-review) so the reviewers, author, and editor(s) are all in the same place and can discuss progress in real time as the author makes any required updates. For specific software issues, it's better to open an issue on the GitHub issue tracker for the specific piece of software.

In this case, we're actually assessing a set of packages (list here: https://github.com/openjournals/joss-reviews/issues/7805#issuecomment-2784703379). They are all installed by the umbrella package and should work interoperably though, so hopefully you should be able to review the whole set in one go.

I'm going to try start the review process now...
