### 2025-04-17 #8066 @njlyon0
Hi @goldingn

I'm reading the COI and may have a minor conflict of interest that I want to run by you to see if I'm still an eligible reviewer. Colin Smith (2nd author) and I work at separate organizations that have a reasonable close association. Similarly, while Colin and I have never formally collaborated on funded work or published together we have discussed software development and potential future collaborations (though we haven't made any concrete plans in that vein as of yet).

I feel that I can still be an impartial reviewer and that I don't technically violate JOSS's COI but this felt enough like an edge case that I wanted to flag it for you and be fully transparent.
