### 2025-04-22 #8066 @njlyon0
Hi @goldingn

I've finished a first pass at the paper/GitHub repo. Broadly, I think this is a strong piece of software that is well-supported by the  manuscript as-is. However, I think there are two small(ish) changes that could be made to great effect.

1. Expand the vignette to include this package's purpose so users know how to engage with it / the packages it contains.

The [`tidyverse` package website](https://tidyverse.tidyverse.org/)--which I think is a fair analog given that the authors make the same comparison--includes a pseudo-vignette that articulates the `tidyverse` philosophy / purpose which would benefit this package. I think potentially linking this paper (or a subset of the text in the 'Statement of Need' section) would definitely accomplish this. Some of this information is already in the README of the package but I think many R users are more accustomed to using a package's vignette to get this kind of information.

2. Add unit tests for the functions included in this package

The `tidyverse` meta-package includes unit tests (see [here](https://github.com/tidyverse/tidyverse/tree/main/tests/testthat)) so I think it would be reasonable for the authors to include similar tests for `NPSdataverse`. The R-CMD check is a really useful baseline but I think writing specific unit tests (ideally with the [`testthat` package](https://testthat.r-lib.org/)) often leads to functions that are written more defensively and are thus more robust / useful to users with a range of coding experience.

Other than those two points I think this package is nicely-composed and would be _really_ useful to the community of researchers who use EML as their preferred metadata language!
