+++ README.Rmd
@@ -47,7 +47,7 @@ NPSdataverse will install the following R packages:
 |---------|---------|
 |**QCkit**| https://nationalparkservice.github.io/QCkit/|
 |EMLassemblyline (EAL)| https://ediorg.github.io/EMLassemblyline/|
-|EML| https://www.cran-e.com/package/EML|
+|EML| https://cran.r-project.org/package=EML|
 |**EMLeditor**| https://nationalparkservice.github.io/EMLeditor/|
 |**DPchecker**| https://nationalparkservice.github.io/DPchecker/|
 |**NPSutils**| https://nationalparkservice.github.io/NPSutils/|
