### 2025-04-22 309af5f
Fix import warnings in R CMD check
