+++ README.Rmd
@@ -45,12 +45,12 @@ NPSdataverse will install the following R packages:
 
 | Package | Location|
 |---------|---------|
-|**QCkit**| https://nationalparkservice.github.io/QCkit/|
-|EMLassemblyline (EAL)| https://ediorg.github.io/EMLassemblyline/|
-|EML| https://www.cran-e.com/package/EML|
-|**EMLeditor**| https://nationalparkservice.github.io/EMLeditor/|
-|**DPchecker**| https://nationalparkservice.github.io/DPchecker/|
-|**NPSutils**| https://nationalparkservice.github.io/NPSutils/|
+|**QCkit**| https://nationalparkservice.github.io/QCkit|
+|EMLassemblyline (EAL)| https://ediorg.github.io/EMLassemblyline|
+|EML| https://cran.r-project.org/package=EML|
+|**EMLeditor**| https://nationalparkservice.github.io/EMLeditor|
+|**DPchecker**| https://nationalparkservice.github.io/DPchecker|
+|**NPSutils**| https://nationalparkservice.github.io/NPSutils|
 
 ***
 (R packages in **bold** are developed by NPS)
