### 2025-04-22 851d686
Update vignette to include excerpts from the JOSS paper and add usage section to readme
