+++ README.Rmd
@@ -55,6 +55,14 @@ NPSdataverse will install the following R packages:
 ***
 (R packages in **bold** are developed by NPS)
 
+## Usage
+
+```{r setup, eval = FALSE}
+pak::pkg_install("nationalparkservice/NPSdataverse")
+library(NPSdataverse)
+```
+
+
 ## Trouble shooting installation
 
 Each package within the NPSdataverse has a number of dependencies, most of which
