### 2025-04-22 #8066 @angelchen7
Hi @goldingn I'm done with an initial review of the `NPSdataverse` package and its associating paper.

Overall, I think it's a wonderful addition to the journal. I agree with @njlyon0 in that adding content to the vignette would enhance this package. In reviewing `NPSdataverse`, I noticed that the README mentioned a [EML creation template](https://github.com/nationalparkservice/NPSdataverse?tab=readme-ov-file#creating-data-packages) that users can follow. The sort of content that is in the template right now is honestly something that I would expect to see in a vignette, although I do understand the rationale in making it a script template. I wonder if the developers can consider expanding the vignette to include content similar to the template (or a summarized version of it?).

I've also noticed some typos in the template and I've created an [issue](https://github.com/nationalparkservice/EMLeditor/issues/196) for it in its respective repo.
