### 2025-04-22 20405df
Fix headers in vignette
