### 2025-05-10 #8066 @stephpenn1
Hi @goldingn, I have completed my review.

I believe this is a novel and interesting software package, particularly in the era of open data, and suitable for JOSS. Data QC is a growing issue as data volumes increase. The `NPSdataverse` package makes it easy to install the suite of packages. I agree with the reviewers above that enhancing the vignette will greatly elevate the functionality.

I am wondering if a tutorial or example dataset could be useful here. I found the documentation and paper description engaging but difficult to visualize the workflow of how to interact with each package. For example, are there specific formats your data columns have to be in for the `QCkit` to work properly? Apologies if I missed this in the documentation and there is one.
