### 2025-05-13 #8066 @goldingn
@RobLBaker & team: great work on the paper and the changes you've made already. The reviews are all complete and positive and the remaining tasks are optional, so it's up to you how much extra work you want to put into the software before publication.

As soon as you are happy with the software, please do the checks in the list above, [create a release](https://docs.github.com/en/repositories/releasing-projects-on-github), [archive on zenodo/figshare](https://help.zenodo.org/docs/github/archive-software/github-upload/), and post the release version and DOI here. As soon as I have those, I'll do my checklist ASAP and we can hit accept! https://joss.readthedocs.io/en/latest/submitting.html#review-process
