### 2025-05-17 #8066 @RobLBaker
Quick question: do we have to archive and mint DOIs on Zenodo/Figshare or can we use any trusted repository? As in, can we archive on DataStore and mint a DOI there (one of our functions, QCkit::create_datastore_script will most of the heavy lifting there and the process is actually modeled after Zenodo's archive and DOI process)?
