### 2025-05-19 c434ba0
update version to 1.1.0 for new release
