### 2025-05-19 f30dae4
Update version numbers and DOIs for NPSdataverse, QCkit, EMLeditor, DPchecker, and NPSutils
