+++ paper.bib
@@ -188,8 +188,8 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   title = {QCkit: NPS Inventory and Monitoring Quality Control Toolkit},
   author = {Baker, Robert and Judd Patterson and Joe DeVivo and Issac Quevedo and Sarah Wright},
   year = {2025},
-  note = {R package version 1.0.0},
-  doi = {10.57830/2307726},
+  note = {R package version 1.1.0},
+  doi = {10.57830/2313106},
   url = {https://github.com/nationalparkservice/QCkit/},
 }
 
@@ -197,7 +197,8 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   title = {NPSdataverse: Tools and Packages for Data and Metadata Manipulation},
   author = {Baker, Robert and Judd Patterson and Joe DeVivo},
   year = {2025},
-  note = {R package version 1.0.0},
+  note = {R package version 1.1.0},
+  doi = {10.57830/2313107},
   url = {https://github.com/nationalparkservice/NPSdataverse},
 }
 
@@ -205,8 +206,8 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   title = {EMLeditor: View and Edit EML Metadata},
   author = {Baker, Robert and Judd Patterson},
   year = {2025},
-  note = {R package version 1.0.0},
-  doi = {10.57830/2307727},
+  note = {R package version 1.1.0},
+  doi = {10.57830/2313108},
   url = {https://github.com/nationalparkservice/EMLeditor},
 }
 
@@ -214,8 +215,8 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   title = {DPchecker: Checks Data Packages for Congruence},
   author = {Baker, Robert and Sarah E. Wright},
   year = {2025},
-  note = {R package version 1.0.0},
-  doi = {10.57830/2307755},
+  note = {R package version 1.1.0},
+  doi = {10.57830/2313109},
   url = {https://nationalparkservice.github.io/DPchecker/},
 }
 
@@ -223,8 +224,8 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   title = {NPSutils: Collection of Functions to read and manipulate information from the NPS DataStore},
   author = {Baker, Robert and Joe DeVivo and Judd Patterson},
   year = {2025},
-  note = {R package version 1.0.0},
-  doi = {10.57830/2307756},
+  note = {R package version 1.1.0},
+  doi = {10.57830/2313110},
   url = {https://github.com/nationalparkservice/NPSutils},
 }
 
