### 2025-05-19 #8066 @RobLBaker
Great!

I have issued new releases for NPSdataverse, QCkit, EMLeditor, DPchecker and NPSutils all now at v1.1.0.

I have archived the source code for each of these packages on DataStore ([irma.nps.gov/datastore](https://irma.nps.gov/DataStore)) and minted DOIs for v1.1.0 of each:

NPSdataverse: https://doi.org/10.57830/2313107
QCkit: https://doi.org/10.57830/2313106
EMLeditor: https://doi.org/10.57830/2313108
DPchecker: https://doi.org/10.57830/2313109
NPSutils: https://doi.org/10.57830/2313110

I also updated the references in paper.bib to reflect the newest version of these packages (1.1.0) and the new DOIs (PR merge pending).
