### 2025-05-20 #8066 @goldingn
See https://github.com/nationalparkservice/NPSdataverse/pull/63 for some minor proof-reading edits to the paper text
