### 2025-05-20 #8066 @goldingn
@RobLBaker editorialbot spotted a missing DOI for the EML R package. Could you please use one instead of the link to ROpenSci docs? This is the one linked to from the GitHub page for that package: [10.5281/zenodo.597560](https://doi.org/10.5281/zenodo.597560)
