### 2025-05-20 #8066 @goldingn
I'm done for this flurry of activity. I think you just need to:
 - accept/debate my PR with minor edits
 - add DOIs tot he paper for the EML and EMLassemblyline packages
 - respond/action re. the archive authors

Then I just need to check whether the DOI/version need updating after you have done those, and then recommend acceptance
