### 2025-05-20 #8066 @RobLBaker
> Thanks for archiving all those [@RobLBaker](https://github.com/RobLBaker)! For the paper, we'll only link to one archive, so the NPSdataverse one should be fine.
>
> I have to "Check that the archive title, author list, version tag, and the license are correct", and in particular check that the authors listed on the archive match the authors of the article. They don't at the moment, only three authors are listed on the archive. Is there a particular reason why, and if not could you please amend that?

The reason is the authors on the manuscript vs. the individual package differ. I added all of the authors for the NPSdataverse R package to the archived code. I can also add the remaining manuscript authors if you would like.
