### 2025-05-20 7baa944
Add DOIs to EML and EMLassemblyline packages
