+++ paper.bib
@@ -173,6 +173,7 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   year = {2024},
   note = {R package version 2.0.6, https://github.com/ropensci/EML/},
   url = {https://docs.ropensci.org/EML/},
+  doi = {10.5281/zenodo.597560},
 }
 
 @Manual{Smith2022,
@@ -182,6 +183,7 @@ url = {https://doi.org/10.1038/sdata.2016.18},
   note = {R package version 3.5.7,
     https://ediorg.github.io/EMLassemblyline/},
   url = {https://github.com/EDIorg/EMLassemblyline},
+  doi = {10.5281/zenodo.2653915},
 }
 
 @Manual{Baker_QCkit2025,
