### 2025-05-21 #8066 @goldingn
Thanks @RobLBaker!

No problem if the authors on MS and individual packages differ, but JOSS needs the archive and paper to have the same authors. At least all paper authors need to be listed in the archive metadata. That's (presumably; I didn't write the rules) because the paper reflects the broader software endeavour and the archive it points to encapsulates the state of the software as described in the paper.

If you can list them all under some metadata field in the archive, we should be good to go.
