### 2025-05-21 #8066 @RobLBaker
Not a problem. I listed all the authors of the manuscript as authors of the archived reference for the the v1.1.0 release of the NPSdatastore R package (I think when we send metadata up to the DOI registry we only send authors and not stewards, contributors, etc so this seemed the best course of action.  You can check to make sure it meets the standards here:

 https://doi.org/10.57830/2313107
