### 2025-05-22 #8066 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
