### 2025-05-22 #8066 @goldingn
Oh damn, I just double checked after hitting that button and realised the archive is under a CC0 license but the software is MIT.

@RobLBaker are you able to change the archive license to MIT? Then the assigned EIC can accept.
Sorry, I should have spotted that yesterday.
