### 2025-05-22 #8066 @RobLBaker
Unfortunately, we don't have that option available, just CC0 or Public Domain. I'll archive at Zenodo instead/in addition where I am sure they do have MIT licensing available.
