+++ paper.md
@@ -18,15 +18,15 @@ tags:
 date: "20 February 2025"
 output: pdf_document
 affiliations:
-- name: National Park Service, USA
+- name: National Park Service, United States of America
   index: 1
-- name: Environmental Data Initiative, USA
+- name: Environmental Data Initiative, United States of America
   index: 2
-- name: University of Wisconsin, USA
+- name: University of Wisconsin, United States of America
   index: 3
-- name: Student Conservation Association, USA
+- name: Student Conservation Association, United States of America
   index: 4
-- name: University of California, Berkeley, USA
+- name: University of California, Berkeley, United States of America
   index: 5
 authors:
 - name: Robert L. Baker
@@ -65,22 +65,21 @@ bibliography: paper.bib
 
 # Summary
 
-The [NPSdataverse](https://nationalparkservice.github.io/NPSdataverse/) is a suite of R packages developed to create, document, publish, and access data and metadata in open and machine-readable formats. NPSdataverse is modeled off of the tidyverse concept of several packages built with a common goal [@Wickham2019]. The NPSdataverse supports Ecological Metadata Language (EML) metadata and .csv data files. Some of the constituent R packages ([EML](https://docs.ropensci.org/EML/) and [EMLassemblyline](https://ediorg.github.io/EMLassemblyline/)) are general-use and aimed at authoring EML documents. Other R packages ([QCkit](https://nationalparkservice.github.io/QCkit/), [EMLeditor](https://nationalparkservice.github.io/EMLeditor/), [DPchecker](https://nationalparkservice.github.io/DPchecker/) and [NPSutils](https://nationalparkservice.github.io/NPSutils/)) are designed and maintained by the National Park Service (NPS). Although many functions within the NPSdataverse packages are NPS-specific (particularly some API calls), whenever possible the functions are written so that they can also be used by the general public. Scientists conducting permitted research in NPS units can utilize the NPSdataverse to efficiently and consistently meet the data delivery requirements of their permits. Additionally, the packages will be useful for data management plans in a wide variety of grant proposals and for anyone that needs to create open data and machine-readable metadata. The ability to swiftly and easily author, edit, and check Ecological Metadata Language (EML) metadata in a reproducible fashion will be useful for data publication at any number of repositories or data journals. Finally, a scripted interface for downloading NPS data and leveraging metadata while loading it into R or other platforms for subsequent analyses and visualizations will be useful to researchers in the government, academia, and industry as well as the public. 
[…]
