### 2025-05-23 #8066 @RobLBaker
I have updated author affiliations in `paper.md` to use full country names and not abbreviations.
