### 2025-05-25 #8066 @goldingn
@RobLBaker @Kevin-Mattheus-Moerman I'm a bit confused about the archive situation now.

The Figshare archive ([zenodo.org/records/15490473](https://zenodo.org/records/15490473)) has an MIT licence, but that lists the DOI [10.57830/2313107](https://doi.org/10.57830/2313107) that resolves to the NPS datastore, which has CC0 licence. This contains the (GitHub) code which has an MIT licence.

We have the NPS (CC0) DOI attached to the paper, so we still have a licence mismatch between code (MIT) and archive (CC0).

@RobLBaker Is there a way for you to re-create a figshare archive directly from the github, with MIT licence and new DOI, rather than this one that links to NPS?
