### 2025-05-27 #8066 @RobLBaker
I have created a second archive on Zenodo with an MIT license that contains a zipped release of NPSdataverse (with a MIT license) and directly links to the GitHub repository, which also has an MIT license:

https://zenodo.org/records/15528554
